Move input into importer direcgtory.
